Google Citations Information

Wraps getting the information from Google Citations and plotting the wordcloud

gcite(author, user, plot_wordcloud = TRUE, author_args = list(),
  title_args = list(), warn = FALSE, force = FALSE, sleeptime = 0,
  ...)

Arguments

author
author name separated by spaces

user
user ID for google Citations

plot_wordcloud
should the wordcloud be plotted

author_args
Arguments to pass to author_cloud

title_args
Arguments to pass to title_cloud

warn
should warnings be printed from wordcloud?

force
If passing a URL and there is a failure, should the program return NULL, passed to gcite_citation_page

sleeptime
time in seconds between http requests, to avoid Google Scholar rate limit

...
additional options passed to gcite_user_info and therefore GET

Value

List from either gcite_user_info or gcite_author_info

Examples

if (!is_travis() & !is_cran()) {
res = gcite(author = "John Muschelli")
paper_df = res$paper_df
gcite_wordcloud(paper_df)
author_cloud(paper_df$authors)
}